+++ game/game_libs/cl_dlls/studio_util.cpp
@@ -23,15 +23,15 @@ void AngleMatrix(const float* angles, float (*matrix)[4])
 	float angle;
 	float sr, sp, sy, cr, cp, cy;
 
-	angle = angles[YAW] * (M_PI * 2 / 360);
-	sy = sin(angle);
-	cy = cos(angle);
-	angle = angles[PITCH] * (M_PI * 2 / 360);
-	sp = sin(angle);
-	cp = cos(angle);
-	angle = angles[ROLL] * (M_PI * 2 / 360);
-	sr = sin(angle);
-	cr = cos(angle);
+	angle = angles[YAW] * (static_cast<float>(M_PI) * 2.0f / 360.0f);
+	sy = sinf(angle);
+	cy = cosf(angle);
+	angle = angles[PITCH] * (static_cast<float>(M_PI) * 2.0f / 360.0f);
+	sp = sinf(angle);
+	cp = cosf(angle);
+	angle = angles[ROLL] * (static_cast<float>(M_PI) * 2.0f / 360.0f);
+	sr = sinf(angle);
+	cr = cosf(angle);
 
 	// matrix = (YAW * PITCH) * ROLL
 	matrix[0][0] = cp * cy;
@@ -127,15 +127,15 @@ void AngleQuaternion(float* angles, vec4_t quaternion)
 	float sr, sp, sy, cr, cp, cy;
 
 	// FIXME: rescale the inputs to 1/2 angle
-	angle = angles[2] * 0.5;
-	sy = sin(angle);
-	cy = cos(angle);
-	angle = angles[1] * 0.5;
-	sp = sin(angle);
-	cp = cos(angle);
-	angle = angles[0] * 0.5;
-	sr = sin(angle);
-	cr = cos(angle);
+	angle = angles[2] * 0.5f;
+	sy = sinf(angle);
+	cy = cosf(angle);
+	angle = angles[1] * 0.5f;
+	sp = sinf(angle);
+	cp = cosf(angle);
+	angle = angles[0] * 0.5f;
+	sr = sinf(angle);
+	cr = cosf(angle);
 
 	quaternion[0] = sr * cp * cy - cr * sp * sy;  // X
 	quaternion[1] = cr * sp * cy + sr * cp * sy;  // Y
@@ -177,14 +177,14 @@ void QuaternionSlerp(vec4_t p, vec4_t q, float t, vec4_t qt)
 	{
 		if ( (1.0 - cosom) > 0.000001 )
 		{
-			omega = acos(cosom);
-			sinom = sin(omega);
-			sclp = sin((1.0 - t) * omega) / sinom;
-			sclq = sin(t * omega) / sinom;
+			omega = acosf(cosom);
+			sinom = sinf(omega);
+			sclp = sinf((1.0f - t) * omega) / sinom;
+			sclq = sinf(t * omega) / sinom;
 		}
 		else
 		{
-			sclp = 1.0 - t;
+			sclp = 1.0f - t;
 			sclq = t;
 		}
 		for ( i = 0; i < 4; i++ )
@@ -198,8 +198,8 @@ void QuaternionSlerp(vec4_t p, vec4_t q, float t, vec4_t qt)
 		qt[1] = q[0];
 		qt[2] = -q[3];
 		qt[3] = q[2];
-		sclp = sin((1.0 - t) * (0.5 * M_PI));
-		sclq = sin(t * (0.5 * M_PI));
+		sclp = sinf((1.0f - t) * (0.5f * static_cast<float>(M_PI)));
+		sclq = sinf(t * (0.5f * static_cast<float>(M_PI)));
 		for ( i = 0; i < 3; i++ )
 		{
 			qt[i] = sclp * p[i] + sclq * qt[i];
@@ -215,17 +215,17 @@ QuaternionMatrix
 */
 void QuaternionMatrix(vec4_t quaternion, float (*matrix)[4])
 {
-	matrix[0][0] = 1.0 - 2.0 * quaternion[1] * quaternion[1] - 2.0 * quaternion[2] * quaternion[2];
-	matrix[1][0] = 2.0 * quaternion[0] * quaternion[1] + 2.0 * quaternion[3] * quaternion[2];
-	matrix[2][0] = 2.0 * quaternion[0] * quaternion[2] - 2.0 * quaternion[3] * quaternion[1];
+	matrix[0][0] = 1.0f - 2.0f * quaternion[1] * quaternion[1] - 2.0f * quaternion[2] * quaternion[2];
+	matrix[1][0] = 2.0f * quaternion[0] * quaternion[1] + 2.0f * quaternion[3] * quaternion[2];
+	matrix[2][0] = 2.0f * quaternion[0] * quaternion[2] - 2.0f * quaternion[3] * quaternion[1];
 
-	matrix[0][1] = 2.0 * quaternion[0] * quaternion[1] - 2.0 * quaternion[3] * quaternion[2];
-	matrix[1][1] = 1.0 - 2.0 * quaternion[0] * quaternion[0] - 2.0 * quaternion[2] * quaternion[2];
-	matrix[2][1] = 2.0 * quaternion[1] * quaternion[2] + 2.0 * quaternion[3] * quaternion[0];
+	matrix[0][1] = 2.0f * quaternion[0] * quaternion[1] - 2.0f * quaternion[3] * quaternion[2];
+	matrix[1][1] = 1.0f - 2.0f * quaternion[0] * quaternion[0] - 2.0f * quaternion[2] * quaternion[2];
+	matrix[2][1] = 2.0f * quaternion[1] * quaternion[2] + 2.0f * quaternion[3] * quaternion[0];
 
-	matrix[0][2] = 2.0 * quaternion[0] * quaternion[2] + 2.0 * quaternion[3] * quaternion[1];
-	matrix[1][2] = 2.0 * quaternion[1] * quaternion[2] - 2.0 * quaternion[3] * quaternion[0];
-	matrix[2][2] = 1.0 - 2.0 * quaternion[0] * quaternion[0] - 2.0 * quaternion[1] * quaternion[1];
+	matrix[0][2] = 2.0f * quaternion[0] * quaternion[2] + 2.0f * quaternion[3] * quaternion[1];
+	matrix[1][2] = 2.0f * quaternion[1] * quaternion[2] - 2.0f * quaternion[3] * quaternion[0];
+	matrix[2][2] = 1.0f - 2.0f * quaternion[0] * quaternion[0] - 2.0f * quaternion[1] * quaternion[1];
 }
 
 /*
