+++ game/game_libs/cl_dlls/text_message.cpp
@@ -132,7 +132,7 @@ const char* CHudTextMessage::LookupString(const char* msg, int* msg_dest)
 
 void StripEndNewlineFromString(char* str)
 {
-	int s = strlen(str) - 1;
+	int s = static_cast<int>(strlen(str)) - 1;
 	if ( str[s] == '\n' || str[s] == '\r' )
 		str[s] = 0;
 }
